Changed Current Limit and Delin Power to doubles, removed some unused dashboard imports. Fixed a couple depreciated methods. Still some depreciated methods left in CANLights

+++ src/org/usfirst/frc2619/SteamBot2017/MathUtil.java
@@ -7,7 +7,7 @@ public class MathUtil {
 	public static final boolean ALLOW_DELINEARIZATION_DEFAULT = true;
 	public static final String ALLOW_DELINEARIZATION_KEY = "Allow_Delinearization";
 
-	public static double delinearize(double input, int power) {
+	public static double delinearize(double input, double power) {
 		double dbY = Robot.driveTrain.deadband_y;
 		double ret;
 		Allow_Delinearization = false;
