+++ src/org/usfirst/frc2619/SteamBot2017/commands/ArcadeDrive.java
@@ -51,7 +51,7 @@ public class ArcadeDrive extends Command {
 	// Called repeatedly when this Command is scheduled to run
 	protected void execute() {
 		double rightSpeed, leftSpeed, dbX = Robot.driveTrain.deadband_x, dbY = Robot.driveTrain.deadband_y;
-		int power = Robot.driveTrain.delin_pow;
+		double power = Robot.driveTrain.delin_pow;
 		leftSpeed = MathUtil.deadbandCheck(-1 * Robot.oi.getLeftJoystick().getY(), dbY)
 				+ MathUtil.deadbandCheck(Robot.oi.getLeftJoystick().getX(), dbX);
 		rightSpeed = MathUtil.deadbandCheck(-1 * Robot.oi.getLeftJoystick().getY(), dbY)
